+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Full Erlang target support alongside JavaScript
 - Add proper optional error handling API
+- Add `intlrelative` to format duration like `Intl.RelativeTimeFormat`
 
 ## [1.0.0] - 25/02/2026